requirements.txt
beautifulsoup4==4.13.3
mysql-connector-python==0.3.1
pandas==2.2.3
Pillow==11.1.0
pyecharts==2.0.8
pyyaml==6.0.2
qwen_vl_utils==0.0.10
requests==2.32.3
rich==14.0.0
selenium==4.31.0
transformers==4.49.0
